requirements.txt
jira==3.8.0
pandas==2.2.0
matplotlib==3.8.2
google-genai==0.3.0
python-dotenv==1.0.1